AWM_CA2_Django/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.6
GDAL==3.4.3
numpy==1.26.0
psycopg2==2.9.6
pyproj==3.4.1
sqlparse==0.4.4
typing_extensions==4.8.0
djangorestframework==3.14.0
django-cors-headers==3.9.0
